requirements.txt
# python 3.13
# PyQt5==5.15.9
# requests==2.31.0
# matplotlib==3.9.2
# reportlab==4.2.5
# openpyxl==3.1.5
# Pillow==11.0.0
# PyPDF2==3.0.1
# python-dotenv==1.0.0
seaborn==0.12.2
# numpy==2.3.2
# pandas==2.3.2

# python 3.11
PyQt5==5.15.11
requests==2.31.0
matplotlib==3.9.2
reportlab==4.2.5
openpyxl==3.1.5
Pillow==10.4.0
PyPDF2==3.0.1
python-dotenv==1.0.0
seaborn==0.13.2
numpy==1.26.4
pandas==2.2.2
